--- internal-api/src/test/groovy/datadog/trace/api/http/HttpResourceNamesTest.groovy
@@ -1,20 +1,0 @@
-package datadog.trace.api.http
-
-import datadog.trace.api.normalize.HttpResourceNames
-import spock.lang.Specification
-
-class HttpResourceNamesTest extends Specification {
-  def "works as expected" () {
-    when:
-    def resourceName = HttpResourceNames.join(method, path)
-
-    then:
-    resourceName.toString() == expected
-
-    where:
-    method | path | expected
-    "GET"  | "/test" | "GET /test"
-    null   | "/test" | "/test"
-    "GET"  | null    | "/"
-  }
-}
