+++ internal-api/src/test/groovy/datadog/trace/api/normalize/AntPathMatcherTest.groovy
@@ -1,5 +1,6 @@
 package datadog.trace.api.normalize
 
+
 import spock.lang.Shared
 import spock.lang.Specification
